Renaming search-service.ts => SearchService.ts
