+++ packages/search/search.ts
@@ -18,4 +18,4 @@
 export * from './src/query';
 export * from './src/solr-query';
 export * from './src/elastic-query';
-export * from './src/search-service';
+export * from './src/SearchService';
